refactor(idx): split idx `shared.ts` file into `shared` folder
The `shared.ts` file whas getting very large and unmanagable with all the different things it was doing.

This PR splits the file up into files inside a `shared` folder, each with an individual task making it easier to reason about.
